+++ www/src/utils/node/docs.js
@@ -1,31 +1,19 @@
-const _ = require(`lodash`)
-const minimatch = require(`minimatch`)
-
 const { getPrevAndNext } = require(`../get-prev-and-next.js`)
 const { getMdxContentSlug } = require(`../get-mdx-content-slug`)
 const { getTemplate } = require(`../get-template`)
-const findApiCalls = require(`../find-api-calls`)
 
-const ignorePatterns = [
-  `**/commonjs/**`,
-  `**/node_modules/**`,
-  `**/__tests__/**`,
-  `**/dist/**`,
-  `**/__mocks__/**`,
-  `babel.config.js`,
-  `graphql.js`,
-  `**/flow-typed/**`,
-]
-
-function isCodeFile(node) {
-  return (
-    node.internal.type === `File` &&
-    node.sourceInstanceName === `gatsby-core` &&
-    [`js`].includes(node.extension) &&
-    !ignorePatterns.some(ignorePattern =>
-      minimatch(node.relativePath, ignorePattern)
-    )
-  )
+function mdxResolverPassthrough(fieldName) {
+  return async (source, args, context, info) => {
+    const type = info.schema.getType(`Mdx`)
+    const mdxNode = context.nodeModel.getNodeById({
+      id: source.parent,
+    })
+    const resolver = type.getFields()[fieldName].resolve
+    const result = await resolver(mdxNode, args, context, {
+      fieldName,
+    })
+    return result
+  }
 }
 
 // convert a string like `/some/long/path/name-of-docs/` to `name-of-docs`
@@ -37,99 +25,88 @@ const slugToAnchor = slug =>
 
 exports.createSchemaCustomization = ({ actions: { createTypes } }) => {
   createTypes(/* GraphQL */ `
-    type Mdx implements Node {
-      frontmatter: MdxFrontmatter
-      fields: MdxFields
-    }
-
-    type MdxFrontmatter @dontInfer {
+    type DocPage implements Node @dontInfer @childOf(types: ["Mdx"]) {
+      slug: String!
+      anchor: String!
+      relativePath: String!
+      # Frontmatter-derived fields
       title: String!
-      description: String
-      contentsHeading: String
-      showTopLevelSignatures: Boolean
+      description: String # TODO this should default to excerpt
       disableTableOfContents: Boolean
       tableOfContentsDepth: Int
       issue: String
+      # Frontmatter fields for API docs
       jsdoc: [String!]
       apiCalls: String
-    }
-
-    type MdxFields @dontInfer {
-      slug: String
-      anchor: String
-      section: String
-      locale: String
-    }
-
-    type GatsbyAPICall implements Node @derivedTypes @dontInfer {
-      name: String
-      file: String
-      group: String
-      codeLocation: GatsbyAPICallCodeLocation
-    }
-
-    type GatsbyAPICallCodeLocation @dontInfer {
-      filename: Boolean
-      end: GatsbyAPICallEndpoint
-      start: GatsbyAPICallEndpoint
-    }
-
-    type GatsbyAPICallEndpoint @dontInfer {
-      column: Int
-      line: Int
+      contentsHeading: String
+      showTopLevelSignatures: Boolean
+      # Fields derived from Mdx
+      body: String!
+      timeToRead: Int
+      tableOfContents: JSON
+      excerpt: String!
     }
   `)
+}
+
+exports.createResolvers = ({ createResolvers }) => {
+  createResolvers({
+    DocPage: {
+      body: {
+        resolve: mdxResolverPassthrough(`body`),
+      },
+      timeToRead: {
+        resolve: mdxResolverPassthrough(`timeToRead`),
+      },
+      tableOfContents: {
+        resolve: mdxResolverPassthrough(`tableOfContents`),
+      },
+      excerpt: {
+        resolve: mdxResolverPassthrough(`excerpt`),
+      },
+    },
+  })
 }
 
 exports.onCreateNode = async ({
   node,
   actions,
   getNode,
-  loadNodeContent,
   createNodeId,
   createContentDigest,
 }) => {
-  const { createNode, createParentChildLink, createNodeField } = actions
-
-  if (isCodeFile(node)) {
-    const calls = await findApiCalls({ node, loadNodeContent })
-    if (calls.length > 0) {
-      calls.forEach(call => {
-        const apiCallNode = {
-          id: createNodeId(`findApiCalls-${JSON.stringify(call)}`),
-          parent: node.id,
-          children: [],
-          ...call,
-          internal: {
-            type: `GatsbyAPICall`,
-          },
-        }
-        apiCallNode.internal.contentDigest = createContentDigest(apiCallNode)
-
-        createNode(apiCallNode)
-        createParentChildLink({ parent: node, child: apiCallNode })
-      })
-    }
-    return
-  }
+  const { createNode, createParentChildLink } = actions
 
   const slug = getMdxContentSlug(node, getNode(node.parent))
   if (!slug) return
 
-  const locale = `en`
+  // const locale = `en`
   const section = slug.split(`/`)[1]
   // fields for blog pages are handled in `utils/node/blog.js`
   if (section === `blog`) return
 
-  // Add slugs and other fields for docs pages
-  if (slug) {
-    createNodeField({ node, name: `anchor`, value: slugToAnchor(slug) })
-    createNodeField({ node, name: `slug`, value: slug })
-    createNodeField({ node, name: `section`, value: section })
+  const fieldData = {
+    ...node.frontmatter,
+    slug,
+    anchor: slugToAnchor(slug),
+    relativePath: getNode(node.parent).relativePath,
   }
-  if (locale) {
-    createNodeField({ node, name: `locale`, value: locale })
-  }
+
+  const docPageId = createNodeId(`${node.id} >>> DocPage`)
+  await createNode({
+    ...fieldData,
+    // Required fields.
+    id: docPageId,
+    parent: node.id,
+    children: [],
+    internal: {
+      type: `DocPage`,
+      contentDigest: createContentDigest(fieldData),
+      content: JSON.stringify(fieldData),
+      description: `A documentation page`,
+    },
+  })
+  createParentChildLink({ parent: node, child: getNode(docPageId) })
 }
 
 exports.createPages = async ({ graphql, actions }) => {
@@ -140,23 +117,12 @@ exports.createPages = async ({ graphql, actions }) => {
 
   const { data, errors } = await graphql(/* GraphQL */ `
     query {
-      allMdx(
-        limit: 10000
-        filter: {
-          fileAbsolutePath: { ne: null }
-          fields: { locale: { eq: "en" }, section: { ne: "blog" } }
-        }
-      ) {
+      allDocPage(limit: 10000) {
         nodes {
-          fields {
-            slug
-            locale
-          }
-          frontmatter {
-            title
-            jsdoc
-            apiCalls
-          }
+          slug
+          title
+          jsdoc
+          apiCalls
         }
       }
     }
@@ -164,32 +130,29 @@ exports.createPages = async ({ graphql, actions }) => {
   if (errors) throw errors
 
   // Create docs pages.
-  data.allMdx.nodes.forEach(node => {
-    const slug = _.get(node, `fields.slug`)
-    const locale = _.get(node, `fields.locale`)
-    if (!slug) return
+  data.allDocPage.nodes.forEach(node => {
+    if (!node.slug) return
 
-    const prevAndNext = getPrevAndNext(node.fields.slug)
-    if (node.frontmatter.jsdoc) {
+    const prevAndNext = getPrevAndNext(node.slug)
+    if (node.jsdoc) {
       // API template
       createPage({
-        path: `${node.fields.slug}`,
+        path: `${node.slug}`,
         component: apiTemplate,
         context: {
-          slug: node.fields.slug,
-          jsdoc: node.frontmatter.jsdoc,
-          apiCalls: node.frontmatter.apiCalls,
+          slug: node.slug,
+          jsdoc: node.jsdoc,
+          apiCalls: node.apiCalls,
           ...prevAndNext,
         },
       })
     } else {
       // Docs template
       createPage({
-        path: `${node.fields.slug}`,
+        path: `${node.slug}`,
         component: docsTemplate,
         context: {
-          slug: node.fields.slug,
-          locale,
+          slug: node.slug,
           ...prevAndNext,
         },
       })