+++ packages/babel-preset-gatsby-package/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.2](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby-package@0.5.1...babel-preset-gatsby-package@0.5.2) (2020-07-09)
+
+**Note:** Version bump only for package babel-preset-gatsby-package
+
 ## [0.5.1](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby-package@0.5.0...babel-preset-gatsby-package@0.5.1) (2020-07-02)
 
 **Note:** Version bump only for package babel-preset-gatsby-package